哼

+++ src/main/java/online/rubick/applications/entity/rubick/Files.java
@@ -20,7 +20,7 @@ public class Files implements Serializable {
 
     private String fileUrl;
 
-    private Integer size;
+    private Long size;
 
     private String relatedId;
 
@@ -102,11 +102,11 @@ public class Files implements Serializable {
         this.fileUrl = fileUrl == null ? null : fileUrl.trim();
     }
 
-    public Integer getSize() {
+    public Long getSize() {
         return size;
     }
 
-    public void setSize(Integer size) {
+    public void setSize(Long size) {
         this.size = size;
     }
